app/data: Update "OpenShift Origin" -> "OKD"
Catching up with openshift/origin@f3cfa8ed (Update README.md to
reflect OKD rebranding, 2018-08-03, openshift/origin#20533).  The URL
changes follow the existing redirects:

  $ date --iso=m --utc
  2018-12-02T07:25+0000
  $ curl -sIL http://www.openshift.org | grep 'HTTP\|Location'
  HTTP/1.1 302 Found
  Location: https://www.openshift.org/
  HTTP/1.1 301 Moved Permanently
  Location: https://www.okd.io/
  HTTP/1.1 200 OK
  $ curl -sIL https://www.openshift.org/minishift/ | grep 'HTTP\|Location'
  HTTP/1.1 301 Moved Permanently
  Location: https://www.okd.io/minishift/
  HTTP/1.1 200 OK

I've included a trailing period in featured.json but not in
projects.json to match the apparent conventions in both files.

The new logo was created from the okd.io SVG (as of
openshift-cs/okd.io@e1b69575, Fix double space, 2018-12-01) with:

  $ git diff -U0
  diff --git a/source/img/okd_logo.svg b/source/img/okd_logo.svg
  index 8e470ce..de29a54 100644
  --- a/source/img/okd_logo.svg
  +++ b/source/img/okd_logo.svg
  @@ -7 +7 @@
  -	 .st1{fill:#FFFFFF;}
  +	 .st1{fill:#000000;}
  $ inkscape --without-gui --export-png=logo-okd.png --export-dpi=200 --export-background-opacity=0 --export-area-drawing source/img/okd_logo.svg

using:

  $ inkscape --version
  Inkscape 0.92.2 (5c3e80d, 2017-08-06)

I guessed at the DPI to get a logo with a height in the vicinity of
the other logos.  Personally, I'd be happier with vector source; the
upstream SVG is smaller (1.8 KB vs. this PNG's 12 KB) and means I
don't have to care how much resolution this site needs.